ch02/pyproject.toml
[project]
name = "ch02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.31.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.6",
    "langchain-community>=0.3.27",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]
